@@ -3,7 +3,7 @@ package com.sldlt.controller;
 import javax.servlet.RequestDispatcher;
 import javax.servlet.http.HttpServletRequest;
 
-import org.springframework.boot.autoconfigure.web.ErrorController;
+import org.springframework.boot.web.servlet.error.ErrorController;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -31,10 +31,5 @@ public class BaseErrorController implements ErrorController {
         } else {
             return "error";
         }
-    }
-
-    @Override
-    public String getErrorPath() {
-        return PATH;
     }
 }
