@@ -10,11 +10,11 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
-@ControllerAdvice
+@ControllerAdvice({ "com.sldlt.downloader.resource", "com.sldlt.navps.resource", "com.sldlt.orders.resource" })
 public class RestResponseEntityExceptionHandler extends ResponseEntityExceptionHandler {
 
     @ExceptionHandler(Exception.class)
-    public ResponseEntity<Object> handleConflict(final Exception ex, final WebRequest request) {
+    public ResponseEntity<Object> handleGenericException(final Exception ex, final WebRequest request) {
         logger.error(ex.getMessage(), ex);
         return handleExceptionInternal(ex, Collections.singletonMap("error", "Something went wrong."), new HttpHeaders(),
             HttpStatus.INTERNAL_SERVER_ERROR, request);
