@@ -1,12 +1,12 @@
 package com.sldlt.downloader.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.querydsl.QueryDslPredicateExecutor;
+import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.stereotype.Component;
 
 import com.sldlt.downloader.entity.Task;
 
 @Component
-public interface TaskRepository extends JpaRepository<Task, Long>, QueryDslPredicateExecutor<Task> {
+public interface TaskRepository extends JpaRepository<Task, Long>, QuerydslPredicateExecutor<Task> {
 
 }
